Passage de int à double pour precalc de read.c
Le precalc calculé lorsqu'il y avait une impasse prenait l'addition de
precalcSurfer+alphaDivN alors qu'il était de type int.

+++ read.c
@@ -110,7 +110,7 @@ int readFile(char * filename, Element*** pIndex, int* pN, uint8_t** pEmptyLines,
 	fclose(f);
 
 	if(flagImpasse){
-		int precalc = alphaDivN + precalcSurfer;
+		double precalc = alphaDivN + precalcSurfer;
 		for(int k = 0; k < n ; k++)
 			max(delta+k,precalc);
 	}
